agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "File scanning agent with schema-first config and versioned outputs"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "jsonschema>=4.22",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
